% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_create.R
\name{bls_create}
\alias{bls_create}
\title{Initializes new BLS object.}
\usage{
bls_create(key = NA)
}
\arguments{
\item{key}{\code{base::character} object which sets API key in \code{BLS} object}

\item{BLS}{\code{bls4R::BLS} object containing URL and API key}
}
\description{
This function creates a BLS object to interact with the BLS functions.
}
\examples{
# Create default BLS object
BLS <- bls4R::bls_create()

# Set API key
BLS <- bls4R::bls_set_key(BLS=BLS,key="YOUR_API_KEY")

# Or, do it all in one call!
BLS <- bls4R::bls_create(key="YOUR_API_KEY")

}
\author{
Issaac Hansen
}
