% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_extract.R
\name{bls_extract}
\alias{bls_extract}
\title{Extracts data from BLS request}
\usage{
bls_extract(request)
}
\arguments{
\item{request}{the data element of the BLS API request from the \code{bls_request()} function}
}
\value{
A \code{data.table::data.table()} containing 3 columns: \code{year}, \code{month}, \code{index}.
}
\description{
This function extracts data from the BLS API request and returns it as a \code{data.table::data.table()}
}
\author{
Issaac Hansen
}
