% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_query.R
\name{bls_query}
\alias{bls_query}
\title{Build BLS API request}
\usage{
bls_query(series, start_year, end_year, url_ = BLS$url)
}
\arguments{
\item{series}{\code{base::character}; the series id corresponding with the BLS coding}

\item{start_year}{\code{base::integer}; the starting year (inclusive) to return BLS indices for}

\item{end_year}{\code{base::integer}; the ending year (inclusive) to return BLS indices for}

\item{url}{\code{base::character}; the url for the BLS api}
}
\value{
A \code{data.table::data.table()} containing 3 columns: \code{year}, \code{month}, and \code{index}
}
\description{
This function returns a formatted url to send an API request
to the BLS website.
}
\author{
Issaac Hansen
}
