% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_request.R
\name{bls_request}
\alias{bls_request}
\title{Request BLS index}
\usage{
bls_request(query)
}
\arguments{
\item{query}{an API request for the BLS website generated by \code{bls_query()}}
}
\value{
A \code{data.table::data.table()} containing 3 columns: \code{year}, \code{month}, and \code{index}
}
\description{
This function returns a number of specified series from the BLS as a
\code{data.table::data.table()} with \code{year}, \code{month}, and \code{index}.
}
\author{
Issaac Hansen
}
