% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_set_key.R
\name{bls_set_key}
\alias{bls_set_key}
\title{Set API key for BLS.}
\usage{
bls_set_key(BLS, key)
}
\arguments{
\item{BLS}{\code{bls4R::BLS} object containing URL and API key}

\item{key}{\code{base::character} object which sets API key in \code{BLS} object}
}
\description{
This function sets the API key in the BLS object to interact with the API.
}
\author{
Issaac Hansen
}
