% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_unpacker.R
\name{bls_unpacker}
\alias{bls_unpacker}
\title{Unpacks a row from the BLS request list}
\usage{
bls_unpacker(x)
}
\arguments{
\item{x}{a row from the list returned by \code{bls_request()} function}
}
\value{
A \code{data.table::data.table()} containing 3 columns: \code{year}, \code{month}, \code{index}.
}
\description{
This function extracts a row of from the list returned from \code{bls_request()}
and returns it as a \code{base::vector()} with the \code{year}, \code{month}, and \code{index}.
}
\author{
Issaac Hansen
}
