% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_requests.R
\name{bls_requests}
\alias{bls_requests}
\title{Request BLS indices}
\usage{
bls_requests(series, start_year, end_year, url_ = BLS$url)
}
\arguments{
\item{series}{\code{base::vector}; a vector of series id's corresponding with the BLS coding}

\item{start_year}{\code{base::integer}; the starting year (inclusive) to return BLS indices for}

\item{end_year}{\code{base::integer}; the ending year (inclusive) to return BLS indices for}

\item{url_}{\code{base::character}; the url for the BLS api}
}
\value{
A \code{data.table::data.table()} containing 4 columns: \code{year}, \code{month}, \code{index}, and \code{series_id}
}
\description{
This function returns a number of specified series from the BLS as a
\code{data.table::data.table()} with \code{year}, \code{month}, \code{index}, and \code{series_id}.
}
\author{
Issaac Hansen
}
